Cloudflare kv cluster store initial edits
