+++ packages/api/package.json
@@ -26,6 +26,7 @@
     "dev:server": "nodemon -w src -x node -r esm src/cli.js -e js,yaml",
     "dev:build-worker": "parcel watch --bundle-node-modules -t browser --out-dir dist-worker src/worker.js",
     "dev:run-worker": "nodemon -w dist-worker/worker.js -x 'cloudworker -p 3011 -d dist-worker/worker.js'",
+    "dev:yaml-json": "nodemon -w src/schema/schema.yaml -x 'js-yaml src/schema/schema.yaml > src/schema/schema.json'",
     "dev": "run-p \"dev:**\"",
     "go-livepeer": "run-p \"go-livepeer:**\"",
     "go-livepeer:broadcaster": "bin/livepeer -broadcaster -datadir ./bin/broadcaster -orchAddr 127.0.0.1:3086 -rtmpAddr 0.0.0.0:3035 -httpAddr :3085 -cliAddr :3075 -v 6 -authWebhookUrl http://localhost:3004/api/stream/hook -orchWebhookUrl http://localhost:3004/api/orchestrator",
@@ -96,6 +97,8 @@
   ],
   "alias": {
     "pg-native": "./src/empty.js",
-    "level": "./src/empty.js"
+    "level": "./src/empty.js",
+    "@kubernetes/client-node": "./src/empty.js",
+    "./src/store/level-store.js": "./src/empty.js"
   }
 }
