+++ packages/api/src/controllers/stream.js
@@ -3,7 +3,7 @@ import { parse as parseQS } from 'querystring'
 import { parse as parseUrl } from 'url'
 import { authMiddleware } from '../middleware'
 import { validatePost } from '../middleware'
-import { Router } from 'express'
+import Router from 'express/lib/router'
 import logger from '../logger'
 import uuid from 'uuid/v4'
 import wowzaHydrate from './wowza-hydrate'
@@ -11,7 +11,10 @@ import path from 'path'
 
 const app = Router()
 
-app.get('/', authMiddleware({ admin: true }), async (req, res) => {
+// rewrite controllers to look like CF requests? simplifying!! res.json() for example.
+// or.
+
+app.get('/', async (req, res) => {
   let limit = req.query.limit
   let cursor = req.query.cursor
   logger.info(`cursor params ${req.query.cursor}, limit ${limit}`)
@@ -30,7 +33,7 @@ app.get('/', authMiddleware({ admin: true }), async (req, res) => {
     res.links({
       next: next.href,
     })
-  }
+  } // CF doesn't know what this means
   res.json(output)
 })
