+++ packages/api/src/schema/schema.json
@@ -1,0 +1,183 @@
+{
+  "openapi": "3.0.1",
+  "info": {
+    "title": "Livepeer API",
+    "version": "0.0.1-alpha.0"
+  },
+  "tags": [
+    {
+      "name": "stream",
+      "description": "Wowza Stream Endpoints"
+    }
+  ],
+  "components": {
+    "schemas": {
+      "stream": {
+        "type": "object",
+        "required": ["id", "kind", "name", "presets"],
+        "additionalProperties": false,
+        "properties": {
+          "id": {
+            "type": "string",
+            "readOnly": true
+          },
+          "kind": {
+            "type": "string",
+            "readOnly": true
+          },
+          "name": {
+            "type": "string"
+          },
+          "presets": {
+            "type": "array",
+            "items": {
+              "type": "string",
+              "enum": [
+                "P720p60fps16x9",
+                "P720p30fps16x9",
+                "P720p30fps4x3",
+                "P576p30fps16x9",
+                "P360p30fps16x9",
+                "P360p30fps4x3",
+                "P240p30fps16x9",
+                "P240p30fps4x3",
+                "P144p30fps16x9"
+              ]
+            }
+          },
+          "wowza": {
+            "type": "object",
+            "required": [
+              "transcoderAppConfig",
+              "transcoderTemplateAppConfig",
+              "streamNameGroups",
+              "sourceInfo"
+            ],
+            "additionalProperties": false,
+            "properties": {
+              "transcoderAppConfig": {
+                "type": "object"
+              },
+              "transcoderTemplateAppConfig": {
+                "type": "object"
+              },
+              "streamNameGroups": {
+                "type": "array"
+              },
+              "sourceInfo": {
+                "type": "object",
+                "required": ["width", "height", "fps"],
+                "additionalProperties": false,
+                "properties": {
+                  "width": {
+                    "type": "integer",
+                    "minValue": 1
+                  },
+                  "height": {
+                    "type": "integer",
+                    "minValue": 1
+                  },
+                  "fps": {
+                    "type": "integer",
+                    "minVallue": 1
+                  }
+                }
+              }
+            }
+          },
+          "renditions": {
+            "type": "object",
+            "additionalProperties": {
+              "type": "string"
+            }
+          }
+        }
+      },
+      "error": {
+        "required": ["errors"],
+        "type": "object",
+        "properties": {
+          "errors": {
+            "type": "array",
+            "minItems": 1,
+            "items": {
+              "type": "string"
+            }
+          }
+        }
+      }
+    }
+  },
+  "paths": {
+    "/stream": {
+      "tags": ["stream"],
+      "post": {
+        "description": "receives Wowza stream",
+        "requestBody": {
+          "required": true,
+          "content": {
+            "application/json": {
+              "schema": {
+                "type": "object",
+                "$ref": "#/components/schemas/stream"
+              }
+            }
+          }
+        },
+        "responses": {
+          "200": {
+            "description": "Success",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "type": "array",
+                  "items": {
+                    "$ref": "#/components/schemas/stream"
+                  }
+                }
+              }
+            }
+          },
+          "default": {
+            "description": "Error",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/error"
+                }
+              }
+            }
+          }
+        }
+      },
+      "get": {
+        "description": "list streams",
+        "responses": {
+          "200": {
+            "description": "Success",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "type": "array",
+                  "items": {
+                    "$ref": "#/components/schemas/stream"
+                  }
+                }
+              }
+            }
+          },
+          "default": {
+            "description": "Error",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/error"
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
